qca-swiss-army-knife: halregdump: add original HAL -> CRDA dump code
This adds the code used to do the final work for the Atheros HAL regulatory
dump to a CRDA db.txt. The Atheros HAL regulatory code was the only source
of information we had at that time for regulatory information.

This code is the final resulting work of the changes made to Atheros
HAL code to extract regulatory information using a new format for CRDA.
The hal-reg-dump tool was used to scrape the Atheros HAL for regulatory
informationa and extract it in a legible form.

Putting this on a public tree for historical purposes, the code itself
diverges quite a bit from what was merged on ath9k on 2.6.27 since this
code was written prior that merge effort.

Signed-off-by: Luis R. Rodriguez <lrodriguez@atheros.com>