Add suport for using the Atheros HAL initvals if you have them
This will allow us to check the checksums on the Atheros HAL
initvals and compare against what we have. The next step will
be to parse these to write them. This will let us dump data in
whatever form we want instead of writing doing silly style fixes
first and then diff'ing them.

Signed-off-by: Luis R. Rodriguez <lrodriguez@atheros.com>